requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pymysql
psycopg2-binary
pyodbc
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow